+++ evaluator.fhir/src/main/java/org/opencds/cqf/cql/evaluator/fhir/util/FhirResourceLoader.java
@@ -1,4 +1,4 @@
-package org.opencds.cqf.cql.evaluator.measure;
+package org.opencds.cqf.cql.evaluator.fhir.util;
 
 
 import ca.uhn.fhir.context.FhirContext;
